requirements.txt
python-telegram-bot
flask
Pillow